requirements.txt
fastapi
uvicorn
pydantic
motor
pandas
streamlit
openai
python-dotenv